7.18号更新
1. 完成订单模块，引入日历插件，实现按时间范围搜索订单。
2. 优化板块管理模块代码
3. 修复完善一些bug

+++ src/com/lexian/manager/goods/service/impl/SortServiceImpl.java
@@ -1,6 +1,8 @@
 package com.lexian.manager.goods.service.impl;
 
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -9,6 +11,7 @@ import com.lexian.manager.goods.bean.Category;
 import com.lexian.manager.goods.dao.SortDao;
 import com.lexian.manager.goods.service.SortService;
 import com.lexian.utils.Constant;
+import com.lexian.web.Page;
 import com.lexian.web.ResultHelper;
 
 @Service
@@ -27,9 +30,17 @@ public class SortServiceImpl implements SortService{
 	}
 
 	@Override
-	public ResultHelper getAllCategories() {
-		List<Category> list = sortDao.getAllCategories();
-		return new ResultHelper(Constant.code_success,list);
+	public ResultHelper getAllCategories(Integer pageNo) {
+		Page page = new Page();
+		page.setPageNo(pageNo);
+		
+		Map<String, Object> params = new HashMap<>();
+		params.put("page", page);
+		List<Category> orderssWithStore = sortDao.getAllCategoriesPage(params);
+		page.setData(orderssWithStore);
+		ResultHelper result = new ResultHelper(Constant.code_success, page);
+
+		return result;
 	}
 	
 	@Override
@@ -60,8 +71,18 @@ public class SortServiceImpl implements SortService{
 
 	@Override
 	public ResultHelper deleteCategory(int id) {
-		sortDao.deleteCategory(id);
-	 return new ResultHelper(Constant.code_success);
+		if (sortDao.getCountCategoryByParentId(id) != 0) {
+			return new ResultHelper(Constant.code_entity_duplicated); 
+		}else{
+			sortDao.deleteCategory(id);
+			return new ResultHelper(Constant.code_success);
+		}
+	}
+
+	@Override
+	public ResultHelper getCategories() {
+		List<Category> list = sortDao.getCategories();
+		return null;
 	}
 
 }
